@@ -6,7 +6,7 @@ All scrapers produce this format, all exporters consume this format.
 
 import json
 from datetime import datetime
-from typing import Any, Dict, List, Optional
+from typing import Any
 
 from pydantic import BaseModel, Field, HttpUrl, field_validator
 
@@ -25,13 +25,13 @@ class CISControl(BaseModel):
 class MITREMapping(BaseModel):
     """MITRE ATT&CK framework mapping."""
 
-    techniques: List[str] = Field(
+    techniques: list[str] = Field(
         default_factory=list, description="MITRE technique IDs (e.g., ['T1068', 'T1203'])"
     )
-    tactics: List[str] = Field(
+    tactics: list[str] = Field(
         default_factory=list, description="MITRE tactic IDs (e.g., ['TA0001'])"
     )
-    mitigations: List[str] = Field(
+    mitigations: list[str] = Field(
         default_factory=list, description="MITRE mitigation IDs (e.g., ['M1022'])"
     )
 
@@ -43,7 +43,7 @@ class Artifact(BaseModel):
     view_level: str = Field(..., description="Artifact reference (e.g., '1.1.1.1.1')")
     title: str
     status: str
-    artifact_type: Dict[str, Any] = Field(..., description="Artifact type metadata")
+    artifact_type: dict[str, Any] = Field(..., description="Artifact type metadata")
 
 
 class ParentReference(BaseModel):
@@ -74,59 +74,59 @@ class Recommendation(BaseModel):
     # ============ Classification & Compliance Mappings ============
     assessment_status: str = Field(..., description="Assessment type: 'Automated' or 'Manual'")
 
-    profiles: List[str] = Field(
+    profiles: list[str] = Field(
         default_factory=list,
         description="Applicable profiles (e.g., ['Level 1 - Server', 'Level 2 - Workstation'])",
     )
 
-    cis_controls: List[CISControl] = Field(
+    cis_controls: list[CISControl] = Field(
         default_factory=list,
         description="CIS Controls v7 and v8 mappings with Implementation Groups",
     )
 
-    mitre_mapping: Optional[MITREMapping] = Field(
+    mitre_mapping: MITREMapping | None = Field(
         None, description="MITRE ATT&CK framework mapping (techniques, tactics, mitigations)"
     )
 
-    nist_controls: List[str] = Field(
+    nist_controls: list[str] = Field(
         default_factory=list, description="NIST SP 800-53 control IDs (e.g., ['SI-3', 'MP-7'])"
     )
 
     # ============ Hierarchical Structure ============
-    parent: Optional[ParentReference] = Field(
+    parent: ParentReference | None = Field(
         None, description="Parent recommendation in hierarchy"
     )
 
-    artifacts: List[Artifact] = Field(
+    artifacts: list[Artifact] = Field(
         default_factory=list, description="Test artifacts for automated compliance checking"
     )
 
     # ============ Content Fields (HTML) ============
-    description: Optional[str] = Field(None, description="Detailed description (HTML format)")
+    description: str | None = Field(None, description="Detailed description (HTML format)")
 
-    rationale: Optional[str] = Field(None, description="Rationale/justification (HTML format)")
+    rationale: str | None = Field(None, description="Rationale/justification (HTML format)")
 
-    impact: Optional[str] = Field(None, description="Impact statement (HTML format)")
+    impact: str | None = Field(None, description="Impact statement (HTML format)")
 
-    audit: Optional[str] = Field(None, description="Audit procedure (HTML format with code blocks)")
+    audit: str | None = Field(None, description="Audit procedure (HTML format with code blocks)")
 
-    remediation: Optional[str] = Field(
+    remediation: str | None = Field(
         None, description="Remediation steps (HTML format with code blocks)"
     )
 
-    additional_info: Optional[str] = Field(
+    additional_info: str | None = Field(
         None, description="Additional information and notes (HTML format)"
     )
 
-    default_value: Optional[str] = Field(
+    default_value: str | None = Field(
         None, description="Default configuration value (HTML format)"
     )
 
-    artifact_equation: Optional[str] = Field(
+    artifact_equation: str | None = Field(
         None, description="Artifact equation logic (HTML format)"
     )
 
-    references: Optional[str] = Field(
+    references: str | None = Field(
         None, description="External references and citations (HTML format)"
     )
 
@@ -145,7 +145,7 @@ class Benchmark(BaseModel):
     scraper_version: str = Field(..., description="Scraper strategy version used")
 
     total_recommendations: int = Field(..., ge=0)
-    recommendations: List[Recommendation] = Field(...)
+    recommendations: list[Recommendation] = Field(...)
 
     @field_validator("total_recommendations")
     @classmethod
