@@ -1,7 +1,6 @@
 from dataclasses import dataclass, field
 from decimal import Decimal
 from enum import Enum
-from typing import Optional, Union
 
 from xsdata.models.datatype import XmlDate, XmlDateTime
 
@@ -33,7 +32,7 @@ class Cpe2IdrefType:
     class Meta:
         name = "CPE2idrefType"
 
-    idref: Optional[str] = field(
+    idref: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -55,14 +54,14 @@ class BenchmarkReferenceType:
     class Meta:
         name = "benchmarkReferenceType"
 
-    href: Optional[str] = field(
+    href: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
             "required": True,
         },
     )
-    id: Optional[str] = field(
+    id: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -103,14 +102,14 @@ class CheckContentRefType:
     class Meta:
         name = "checkContentRefType"
 
-    href: Optional[str] = field(
+    href: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
             "required": True,
         },
     )
-    name: Optional[str] = field(
+    name: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -166,7 +165,7 @@ class CheckExportType:
     class Meta:
         name = "checkExportType"
 
-    value_id: Optional[str] = field(
+    value_id: str | None = field(
         default=None,
         metadata={
             "name": "value-id",
@@ -174,7 +173,7 @@ class CheckExportType:
             "required": True,
         },
     )
-    export_name: Optional[str] = field(
+    export_name: str | None = field(
         default=None,
         metadata={
             "name": "export-name",
@@ -211,7 +210,7 @@ class CheckImportType:
     class Meta:
         name = "checkImportType"
 
-    import_name: Optional[str] = field(
+    import_name: str | None = field(
         default=None,
         metadata={
             "name": "import-name",
@@ -219,7 +218,7 @@ class CheckImportType:
             "required": True,
         },
     )
-    import_xpath: Optional[str] = field(
+    import_xpath: str | None = field(
         default=None,
         metadata={
             "name": "import-xpath",
@@ -341,7 +340,7 @@ class IdentType:
             "required": True,
         },
     )
-    system: Optional[str] = field(
+    system: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -382,14 +381,14 @@ class IdentityType:
             "required": True,
         },
     )
-    authenticated: Optional[bool] = field(
+    authenticated: bool | None = field(
         default=None,
         metadata={
             "type": "Attribute",
             "required": True,
         },
     )
-    privileged: Optional[bool] = field(
+    privileged: bool | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -430,7 +429,7 @@ class IdrefType:
     class Meta:
         name = "idrefType"
 
-    idref: Optional[str] = field(
+    idref: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -496,7 +495,7 @@ class InstanceResultType:
             "type": "Attribute",
         },
     )
-    parent_context: Optional[str] = field(
+    parent_context: str | None = field(
         default=None,
         metadata={
             "name": "parentContext",
@@ -598,7 +597,7 @@ class ParamType:
             "required": True,
         },
     )
-    name: Optional[str] = field(
+    name: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -629,7 +628,7 @@ class PlainTextType:
             "required": True,
         },
     )
-    id: Optional[str] = field(
+    id: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -661,7 +660,7 @@ class ProfileSetValueType:
             "required": True,
         },
     )
-    idref: Optional[str] = field(
+    idref: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -709,13 +708,13 @@ class ReferenceType:
     class Meta:
         name = "referenceType"
 
-    href: Optional[str] = field(
+    href: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
         },
     )
-    override: Optional[bool] = field(
+    override: bool | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -818,19 +817,19 @@ class ScoreType:
     class Meta:
         name = "scoreType"
 
-    value: Optional[Decimal] = field(
+    value: Decimal | None = field(
         default=None,
         metadata={
             "required": True,
         },
     )
-    system: Optional[str] = field(
+    system: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
         },
     )
-    maximum: Optional[Decimal] = field(
+    maximum: Decimal | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -858,7 +857,7 @@ class SelNumType:
     class Meta:
         name = "selNumType"
 
-    value: Optional[Decimal] = field(
+    value: Decimal | None = field(
         default=None,
         metadata={
             "required": True,
@@ -952,7 +951,7 @@ class SignatureType:
     class Meta:
         name = "signatureType"
 
-    w3_org_2000_09_xmldsig_element: Optional[object] = field(
+    w3_org_2000_09_xmldsig_element: object | None = field(
         default=None,
         metadata={
             "type": "Wildcard",
@@ -1023,28 +1022,28 @@ class TailoringReferenceType:
     class Meta:
         name = "tailoringReferenceType"
 
-    href: Optional[str] = field(
+    href: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
             "required": True,
         },
     )
-    id: Optional[str] = field(
+    id: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
             "required": True,
         },
     )
-    version: Optional[str] = field(
+    version: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
             "required": True,
         },
     )
-    time: Optional[XmlDateTime] = field(
+    time: XmlDateTime | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -1072,7 +1071,7 @@ class TailoringVersionType:
             "required": True,
         },
     )
-    time: Optional[XmlDateTime] = field(
+    time: XmlDateTime | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -1103,21 +1102,21 @@ class TargetIdRefType:
     class Meta:
         name = "targetIdRefType"
 
-    system: Optional[str] = field(
+    system: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
             "required": True,
         },
     )
-    href: Optional[str] = field(
+    href: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
             "required": True,
         },
     )
-    name: Optional[str] = field(
+    name: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -1136,7 +1135,7 @@ class UriRefType:
     class Meta:
         name = "uriRefType"
 
-    uri: Optional[str] = field(
+    uri: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -1210,13 +1209,13 @@ class VersionType:
             "required": True,
         },
     )
-    time: Optional[XmlDateTime] = field(
+    time: XmlDateTime | None = field(
         default=None,
         metadata={
             "type": "Attribute",
         },
     )
-    update: Optional[str] = field(
+    update: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -1355,7 +1354,7 @@ class CheckType:
             "namespace": "",
         },
     )
-    check_content: Optional[CheckContentType] = field(
+    check_content: CheckContentType | None = field(
         default=None,
         metadata={
             "name": "check-content",
@@ -1363,7 +1362,7 @@ class CheckType:
             "namespace": "",
         },
     )
-    system: Optional[str] = field(
+    system: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -1376,7 +1375,7 @@ class CheckType:
             "type": "Attribute",
         },
     )
-    id: Optional[str] = field(
+    id: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -1395,7 +1394,7 @@ class CheckType:
             "type": "Attribute",
         },
     )
-    base: Optional[str] = field(
+    base: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -1427,7 +1426,7 @@ class FactType:
             "required": True,
         },
     )
-    name: Optional[str] = field(
+    name: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -1456,7 +1455,7 @@ class HtmlTextType:
     class Meta:
         name = "htmlTextType"
 
-    lang: Optional[Union[str, LangValue]] = field(
+    lang: str | LangValue | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -1500,7 +1499,7 @@ class MessageType:
             "required": True,
         },
     )
-    severity: Optional[MsgSevEnumType] = field(
+    severity: MsgSevEnumType | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -1538,7 +1537,7 @@ class Model:
             "namespace": "",
         },
     )
-    system: Optional[str] = field(
+    system: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -1567,20 +1566,20 @@ class NoticeType:
     class Meta:
         name = "noticeType"
 
-    id: Optional[str] = field(
+    id: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
         },
     )
-    base: Optional[str] = field(
+    base: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
             "namespace": "http://www.w3.org/XML/1998/namespace",
         },
     )
-    lang: Optional[Union[str, LangValue]] = field(
+    lang: str | LangValue | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -1637,7 +1636,7 @@ class ProfileSetComplexValueType(ComplexValueType):
     class Meta:
         name = "profileSetComplexValueType"
 
-    idref: Optional[str] = field(
+    idref: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -1690,7 +1689,7 @@ class SelChoicesType:
             "namespace": "",
         },
     )
-    must_match: Optional[bool] = field(
+    must_match: bool | None = field(
         default=None,
         metadata={
             "name": "mustMatch",
@@ -1764,13 +1763,13 @@ class Status:
         name = "status"
         namespace = "http://checklists.nist.gov/xccdf/1.2"
 
-    value: Optional[StatusType] = field(
+    value: StatusType | None = field(
         default=None,
         metadata={
             "required": True,
         },
     )
-    date: Optional[XmlDate] = field(
+    date: XmlDate | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -1822,7 +1821,7 @@ class TailoringBenchmarkReferenceType(BenchmarkReferenceType):
     class Meta:
         name = "tailoringBenchmarkReferenceType"
 
-    version: Optional[str] = field(
+    version: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -1850,7 +1849,7 @@ class TextType:
             "required": True,
         },
     )
-    lang: Optional[Union[str, LangValue]] = field(
+    lang: str | LangValue | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -1907,7 +1906,7 @@ class ComplexCheckType:
             "namespace": "",
         },
     )
-    operator: Optional[CcOperatorEnumType] = field(
+    operator: CcOperatorEnumType | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -1963,7 +1962,7 @@ class FixType:
     class Meta:
         name = "fixType"
 
-    id: Optional[str] = field(
+    id: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -1993,13 +1992,13 @@ class FixType:
             "type": "Attribute",
         },
     )
-    system: Optional[str] = field(
+    system: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
         },
     )
-    platform: Optional[str] = field(
+    platform: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -2040,7 +2039,7 @@ class HtmlTextWithSubType:
     class Meta:
         name = "htmlTextWithSubType"
 
-    lang: Optional[Union[str, LangValue]] = field(
+    lang: str | LangValue | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -2098,7 +2097,7 @@ class OverrideType:
     class Meta:
         name = "overrideType"
 
-    old_result: Optional[ResultEnumType] = field(
+    old_result: ResultEnumType | None = field(
         default=None,
         metadata={
             "name": "old-result",
@@ -2107,7 +2106,7 @@ class OverrideType:
             "required": True,
         },
     )
-    new_result: Optional[ResultEnumType] = field(
+    new_result: ResultEnumType | None = field(
         default=None,
         metadata={
             "name": "new-result",
@@ -2116,7 +2115,7 @@ class OverrideType:
             "required": True,
         },
     )
-    remark: Optional[TextType] = field(
+    remark: TextType | None = field(
         default=None,
         metadata={
             "type": "Element",
@@ -2124,14 +2123,14 @@ class OverrideType:
             "required": True,
         },
     )
-    time: Optional[XmlDateTime] = field(
+    time: XmlDateTime | None = field(
         default=None,
         metadata={
             "type": "Attribute",
             "required": True,
         },
     )
-    authority: Optional[str] = field(
+    authority: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -2161,14 +2160,14 @@ class ProfileNoteType:
     class Meta:
         name = "profileNoteType"
 
-    lang: Optional[Union[str, LangValue]] = field(
+    lang: str | LangValue | None = field(
         default=None,
         metadata={
             "type": "Attribute",
             "namespace": "http://www.w3.org/XML/1998/namespace",
         },
     )
-    tag: Optional[str] = field(
+    tag: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -2235,14 +2234,14 @@ class ProfileRefineRuleType:
             "namespace": "",
         },
     )
-    idref: Optional[str] = field(
+    idref: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
             "required": True,
         },
     )
-    weight: Optional[Decimal] = field(
+    weight: Decimal | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -2250,19 +2249,19 @@ class ProfileRefineRuleType:
             "total_digits": 3,
         },
     )
-    selector: Optional[str] = field(
+    selector: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
         },
     )
-    severity: Optional[SeverityEnumType] = field(
+    severity: SeverityEnumType | None = field(
         default=None,
         metadata={
             "type": "Attribute",
         },
     )
-    role: Optional[RoleEnumType] = field(
+    role: RoleEnumType | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -2313,20 +2312,20 @@ class ProfileRefineValueType:
             "namespace": "",
         },
     )
-    idref: Optional[str] = field(
+    idref: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
             "required": True,
         },
     )
-    selector: Optional[str] = field(
+    selector: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
         },
     )
-    operator: Optional[ValueOperatorType] = field(
+    operator: ValueOperatorType | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -2362,14 +2361,14 @@ class ProfileSelectType:
             "namespace": "",
         },
     )
-    idref: Optional[str] = field(
+    idref: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
             "required": True,
         },
     )
-    selected: Optional[bool] = field(
+    selected: bool | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -2419,7 +2418,7 @@ class TextWithSubType:
     class Meta:
         name = "textWithSubType"
 
-    lang: Optional[Union[str, LangValue]] = field(
+    lang: str | LangValue | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -2474,7 +2473,7 @@ class FixTextType(HtmlTextWithSubType):
     class Meta:
         name = "fixTextType"
 
-    fixref: Optional[str] = field(
+    fixref: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -2578,7 +2577,7 @@ class ProfileType:
             "namespace": "",
         },
     )
-    version: Optional[VersionType] = field(
+    version: VersionType | None = field(
         default=None,
         metadata={
             "type": "Element",
@@ -2660,14 +2659,14 @@ class ProfileType:
             "namespace": "",
         },
     )
-    signature: Optional[SignatureType] = field(
+    signature: SignatureType | None = field(
         default=None,
         metadata={
             "type": "Element",
             "namespace": "",
         },
     )
-    id: Optional[str] = field(
+    id: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -2688,27 +2687,27 @@ class ProfileType:
             "type": "Attribute",
         },
     )
-    note_tag: Optional[str] = field(
+    note_tag: str | None = field(
         default=None,
         metadata={
             "name": "note-tag",
             "type": "Attribute",
         },
     )
-    extends: Optional[str] = field(
+    extends: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
         },
     )
-    base: Optional[str] = field(
+    base: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
             "namespace": "http://www.w3.org/XML/1998/namespace",
         },
     )
-    id_attribute: Optional[str] = field(
+    id_attribute: str | None = field(
         default=None,
         metadata={
             "name": "Id",
@@ -2775,7 +2774,7 @@ class RuleResultType:
     class Meta:
         name = "ruleResultType"
 
-    result: Optional[ResultEnumType] = field(
+    result: ResultEnumType | None = field(
         default=None,
         metadata={
             "type": "Element",
@@ -2832,7 +2831,7 @@ class RuleResultType:
             "namespace": "",
         },
     )
-    complex_check: Optional[ComplexCheckType] = field(
+    complex_check: ComplexCheckType | None = field(
         default=None,
         metadata={
             "name": "complex-check",
@@ -2840,38 +2839,38 @@ class RuleResultType:
             "namespace": "",
         },
     )
-    idref: Optional[str] = field(
+    idref: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
             "required": True,
         },
     )
-    role: Optional[RoleEnumType] = field(
+    role: RoleEnumType | None = field(
         default=None,
         metadata={
             "type": "Attribute",
         },
     )
-    severity: Optional[SeverityEnumType] = field(
+    severity: SeverityEnumType | None = field(
         default=None,
         metadata={
             "type": "Attribute",
         },
     )
-    time: Optional[XmlDateTime] = field(
+    time: XmlDateTime | None = field(
         default=None,
         metadata={
             "type": "Attribute",
         },
     )
-    version: Optional[str] = field(
+    version: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
         },
     )
-    weight: Optional[Decimal] = field(
+    weight: Decimal | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -3007,7 +3006,7 @@ class ItemType:
             "namespace": "",
         },
     )
-    version: Optional[VersionType] = field(
+    version: VersionType | None = field(
         default=None,
         metadata={
             "type": "Element",
@@ -3062,14 +3061,14 @@ class ItemType:
             "type": "Attribute",
         },
     )
-    cluster_id: Optional[str] = field(
+    cluster_id: str | None = field(
         default=None,
         metadata={
             "name": "cluster-id",
             "type": "Attribute",
         },
     )
-    extends: Optional[str] = field(
+    extends: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -3088,21 +3087,21 @@ class ItemType:
             "type": "Attribute",
         },
     )
-    lang: Optional[Union[str, LangValue]] = field(
+    lang: str | LangValue | None = field(
         default=None,
         metadata={
             "type": "Attribute",
             "namespace": "http://www.w3.org/XML/1998/namespace",
         },
     )
-    base: Optional[str] = field(
+    base: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
             "namespace": "http://www.w3.org/XML/1998/namespace",
         },
     )
-    id: Optional[str] = field(
+    id: str | None = field(
         default=None,
         metadata={
             "name": "Id",
@@ -3213,14 +3212,14 @@ class TestResultType:
     class Meta:
         name = "testResultType"
 
-    benchmark: Optional[BenchmarkReferenceType] = field(
+    benchmark: BenchmarkReferenceType | None = field(
         default=None,
         metadata={
             "type": "Element",
             "namespace": "",
         },
     )
-    tailoring_file: Optional[TailoringReferenceType] = field(
+    tailoring_file: TailoringReferenceType | None = field(
         default=None,
         metadata={
             "name": "tailoring-file",
@@ -3249,14 +3248,14 @@ class TestResultType:
             "namespace": "",
         },
     )
-    identity: Optional[IdentityType] = field(
+    identity: IdentityType | None = field(
         default=None,
         metadata={
             "type": "Element",
             "namespace": "",
         },
     )
-    profile: Optional[IdrefType] = field(
+    profile: IdrefType | None = field(
         default=None,
         metadata={
             "type": "Element",
@@ -3279,7 +3278,7 @@ class TestResultType:
             "namespace": "",
         },
     )
-    target_facts: Optional[TargetFactsType] = field(
+    target_facts: TargetFactsType | None = field(
         default=None,
         metadata={
             "name": "target-facts",
@@ -3348,14 +3347,14 @@ class TestResultType:
             "namespace": "",
         },
     )
-    signature: Optional[SignatureType] = field(
+    signature: SignatureType | None = field(
         default=None,
         metadata={
             "type": "Element",
             "namespace": "",
         },
     )
-    id: Optional[str] = field(
+    id: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -3363,14 +3362,14 @@ class TestResultType:
             "pattern": r"xccdf_[^_]+_testresult_.+",
         },
     )
-    start_time: Optional[XmlDateTime] = field(
+    start_time: XmlDateTime | None = field(
         default=None,
         metadata={
             "name": "start-time",
             "type": "Attribute",
         },
     )
-    end_time: Optional[XmlDateTime] = field(
+    end_time: XmlDateTime | None = field(
         default=None,
         metadata={
             "name": "end-time",
@@ -3378,20 +3377,20 @@ class TestResultType:
             "required": True,
         },
     )
-    test_system: Optional[str] = field(
+    test_system: str | None = field(
         default=None,
         metadata={
             "name": "test-system",
             "type": "Attribute",
         },
     )
-    version: Optional[str] = field(
+    version: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
         },
     )
-    id_attribute: Optional[str] = field(
+    id_attribute: str | None = field(
         default=None,
         metadata={
             "name": "Id",
@@ -3571,7 +3570,7 @@ class TailoringType:
     class Meta:
         name = "tailoringType"
 
-    benchmark: Optional[TailoringBenchmarkReferenceType] = field(
+    benchmark: TailoringBenchmarkReferenceType | None = field(
         default=None,
         metadata={
             "type": "Element",
@@ -3593,7 +3592,7 @@ class TailoringType:
             "namespace": "",
         },
     )
-    version: Optional[TailoringVersionType] = field(
+    version: TailoringVersionType | None = field(
         default=None,
         metadata={
             "type": "Element",
@@ -3617,14 +3616,14 @@ class TailoringType:
             "min_occurs": 1,
         },
     )
-    signature: Optional[SignatureType] = field(
+    signature: SignatureType | None = field(
         default=None,
         metadata={
             "type": "Element",
             "namespace": "",
         },
     )
-    id: Optional[str] = field(
+    id: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -3632,7 +3631,7 @@ class TailoringType:
             "pattern": r"xccdf_[^_]+_tailoring_.+",
         },
     )
-    id_attribute: Optional[str] = field(
+    id_attribute: str | None = field(
         default=None,
         metadata={
             "name": "Id",
@@ -3808,14 +3807,14 @@ class ValueType(ItemType):
             "namespace": "",
         },
     )
-    signature: Optional[SignatureType] = field(
+    signature: SignatureType | None = field(
         default=None,
         metadata={
             "type": "Element",
             "namespace": "",
         },
     )
-    id: Optional[str] = field(
+    id: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -3842,7 +3841,7 @@ class ValueType(ItemType):
             "type": "Attribute",
         },
     )
-    interface_hint: Optional[InterfaceHintType] = field(
+    interface_hint: InterfaceHintType | None = field(
         default=None,
         metadata={
             "name": "interfaceHint",
@@ -3953,7 +3952,7 @@ class RuleType(SelectableItemType):
             "namespace": "",
         },
     )
-    impact_metric: Optional[str] = field(
+    impact_metric: str | None = field(
         default=None,
         metadata={
             "name": "impact-metric",
@@ -3990,7 +3989,7 @@ class RuleType(SelectableItemType):
             "namespace": "",
         },
     )
-    complex_check: Optional[ComplexCheckType] = field(
+    complex_check: ComplexCheckType | None = field(
         default=None,
         metadata={
             "name": "complex-check",
@@ -3998,14 +3997,14 @@ class RuleType(SelectableItemType):
             "namespace": "",
         },
     )
-    signature: Optional[SignatureType] = field(
+    signature: SignatureType | None = field(
         default=None,
         metadata={
             "type": "Element",
             "namespace": "",
         },
     )
-    id: Optional[str] = field(
+    id: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -4097,14 +4096,14 @@ class GroupType(SelectableItemType):
             "namespace": "http://checklists.nist.gov/xccdf/1.2",
         },
     )
-    signature: Optional[SignatureType] = field(
+    signature: SignatureType | None = field(
         default=None,
         metadata={
             "type": "Element",
             "namespace": "",
         },
     )
-    id: Optional[str] = field(
+    id: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -4280,7 +4279,7 @@ class Benchmark:
             "namespace": "",
         },
     )
-    platform_specification: Optional[str] = field(
+    platform_specification: str | None = field(
         default=None,
         metadata={
             "name": "platform-specification",
@@ -4295,7 +4294,7 @@ class Benchmark:
             "namespace": "",
         },
     )
-    version: Optional[VersionType] = field(
+    version: VersionType | None = field(
         default=None,
         metadata={
             "type": "Element",
@@ -4351,14 +4350,14 @@ class Benchmark:
             "type": "Element",
         },
     )
-    signature: Optional[SignatureType] = field(
+    signature: SignatureType | None = field(
         default=None,
         metadata={
             "type": "Element",
             "namespace": "",
         },
     )
-    id: Optional[str] = field(
+    id: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
@@ -4366,7 +4365,7 @@ class Benchmark:
             "pattern": r"xccdf_[^_]+_benchmark_.+",
         },
     )
-    id_attribute: Optional[str] = field(
+    id_attribute: str | None = field(
         default=None,
         metadata={
             "name": "Id",
@@ -4379,20 +4378,20 @@ class Benchmark:
             "type": "Attribute",
         },
     )
-    style: Optional[str] = field(
+    style: str | None = field(
         default=None,
         metadata={
             "type": "Attribute",
         },
     )
-    style_href: Optional[str] = field(
+    style_href: str | None = field(
         default=None,
         metadata={
             "name": "style-href",
             "type": "Attribute",
         },
     )
-    lang: Optional[Union[str, LangValue]] = field(
+    lang: str | LangValue | None = field(
         default=None,
         metadata={
             "type": "Attribute",
